+++ deepinv/optim/optim_iterators/drs.py
@@ -105,7 +105,7 @@ class fStepDRS(fStep):
         else:
             p = z
         return cur_data_fidelity.prox(
-            p, y, physics, cur_params["lambda"] * cur_params["stepsize"]
+            p, y, physics, gamma=cur_params["lambda"] * cur_params["stepsize"]
         )
 
 
@@ -130,4 +130,4 @@ class gStepDRS(gStep):
             p = z
         else:
             p = 2 * x - z
-        return cur_prior.prox(p, cur_params["stepsize"], cur_params["g_param"])
+        return cur_prior.prox(p, cur_params["g_param"], gamma=cur_params["stepsize"])